+++ src/main/java/mapper/MyPageMemberMapper.java
@@ -4,9 +4,9 @@ import domain.MemberVO;
 
 public interface MyPageMemberMapper {
 
-    public MemberVO read(Long memno);
+    public MemberVO read(int memno);
 
-    public int delete(Long memno);
+    public int delete(int memno);
 
     public int update(MemberVO member);
